requirements.txt
openai
langchain
faster-whisper
pyttsx3
gtts
sounddevice
scipy
numpy
streamlit
PyPDF2
chardet